lint
feat: support admin/config/set for minioJob
