+++ pkg/utils/miniojob/minioJob_test.go
@@ -143,25 +143,139 @@ func TestParser(t *testing.T) {
 	for _, tc := range testCase {
 		cmd, err := tc.command(args)
 		if tc.expectError && err == nil {
-			t.Fatalf("expect error")
+			t.Fatalf("expectCommand error")
 		}
 		if !tc.expectError && err != nil {
-			t.Fatalf("expect not a error")
+			t.Fatalf("expectCommand not a error")
 		}
 		if !tc.expectError {
 			if tc.expect.Command != "" && cmd.Command != tc.expect.Command {
-				t.Fatalf("expect %s, but got %s", tc.expect, cmd.Command)
+				t.Fatalf("expectCommand %s, but got %s", tc.expect.Command, cmd.Command)
 			}
 			if tc.expect.FileName != "" {
 				if tc.expect.FileContext != cmd.FileContext {
-					t.Fatalf("expect %s, but got %s", tc.expect.FileContext, cmd.FileContext)
+					t.Fatalf("expectCommand %s, but got %s", tc.expect.FileContext, cmd.FileContext)
 				}
 				if tc.expect.FileExt != cmd.FileExt {
-					t.Fatalf("expect %s, but got %s", tc.expect.FileExt, cmd.FileExt)
+					t.Fatalf("expectCommand %s, but got %s", tc.expect.FileExt, cmd.FileExt)
 				}
 				if tc.expect.FileName != cmd.FileName {
-					t.Fatalf("expect %s, but got %s", tc.expect.FileName, cmd.FileName)
+					t.Fatalf("expectCommand %s, but got %s", tc.expect.FileName, cmd.FileName)
 				}
+			}
+		}
+	}
+}
+
+func TestAdminPolicyCreate(t *testing.T) {
+	mcCommand := "admin/policy/create"
+	funcs := JobOperation[mcCommand]
+	testCase := []struct {
+		name             string
+		args             map[string]string
+		expectError      bool
+		expectCommand    string
+		expectFileNumber int
+	}{
+		{
+			name: "testFull",
+			args: map[string]string{
+				"name":   "mypolicy",
+				"policy": "JsonContent",
+			},
+			expectCommand:    "myminio mypolicy /temp/policy.json",
+			expectFileNumber: 1,
+		},
+		{
+			name: "testError1",
+			args: map[string]string{
+				"name": "mypolicy",
+			},
+			expectCommand: "",
+			expectError:   true,
+		},
+		{
+			name: "testError2",
+			args: map[string]string{
+				"policy": "JsonContent",
+			},
+			expectCommand: "",
+			expectError:   true,
+		},
+	}
+	for _, tc := range testCase {
+		command, err := GenerateMinIOIntervalJobCommand(mcCommand, 0, nil, "test", tc.args, funcs)
+		if !tc.expectError {
+			if err != nil {
+				t.Fatal(err)
+			}
+			if command.Command != tc.expectCommand {
+				t.Fatalf("[%s] expectCommand %s, but got %s", tc.name, tc.expectCommand, command.Command)
+			}
+		} else {
+			if err == nil {
+				t.Fatalf("[%s] expectCommand error", tc.name)
+			}
+		}
+	}
+}
+
+func TestMCConfigSet(t *testing.T) {
+	mcCommand := "admin/config/set"
+	funcs := JobOperation[mcCommand]
+	testCase := []struct {
+		name             string
+		args             map[string]string
+		expectCommand    string
+		expectError      bool
+		expectFileNumber int
+	}{
+		{
+			name: "testFull",
+			args: map[string]string{
+				"webhookName": "webhook1",
+				"endpoint":    "endpoint1",
+				"auth_token":  "token1",
+				"client_cert": "cert1",
+				"client_key":  "key1",
+			},
+			expectCommand:    "myminio webhook1 endpoint=\"endpoint1\" auth_token=\"token1\" client_key=\"/temp/client_key.key\" client_cert=\"/temp/client_cert.pem\"",
+			expectFileNumber: 2,
+		},
+		{
+			name: "testOptionFile",
+			args: map[string]string{
+				"webhookName": "webhook1",
+				"endpoint":    "endpoint1",
+				"auth_token":  "token1",
+				"client_key":  "key1",
+			},
+			expectCommand:    "myminio webhook1 endpoint=\"endpoint1\" auth_token=\"token1\" client_key=\"/temp/client_key.key\"",
+			expectFileNumber: 1,
+		},
+		{
+			name: "testOptionKeyValue",
+			args: map[string]string{
+				"webhookName": "webhook1",
+				"endpoint":    "endpoint1",
+				"client_key":  "key1",
+			},
+			expectCommand:    "myminio webhook1 endpoint=\"endpoint1\" client_key=\"/temp/client_key.key\"",
+			expectFileNumber: 1,
+		},
+	}
+	for _, tc := range testCase {
+		command, err := GenerateMinIOIntervalJobCommand(mcCommand, 0, nil, "test", tc.args, funcs)
+		if !tc.expectError {
+			if err != nil {
+				t.Fatal(err)
+			}
+			if command.Command != tc.expectCommand {
+				t.Fatalf("[%s] expectCommand %s, but got %s", tc.name, tc.expectCommand, command.Command)
+			}
+		} else {
+			if err == nil {
+				t.Fatalf("[%s] expectCommand error", tc.name)
 			}
 		}
 	}
